gcp/requirements.txt
tensorflow==2.13.1
google-cloud-storage==2.14.0
Pillow==9.3.0